3LA 리팩토링 완료 컨트롤러 / DB / 미들웨어 로 분리 각각 기능들 작동되는지 점검 완료
