+++ homework/jinseo_kim/backend/constants/env.constant.js
@@ -1,0 +1,2 @@
+// 시크릿 키
+export const secretKey = '1a2b3c4d';
