+++ homework/jinseo_kim/backend/package.json
@@ -4,7 +4,8 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
-    "nodemon": "^3.1.2"
+    "nodemon": "^3.1.2",
+    "prettier": "^3.3.2"
   },
   "name": "backend",
   "version": "1.0.0",
